Remove some trailing spaces and other useless things

+++ MyApplication2/src/main/java/com/example/myapplication2/Animator.java
@@ -45,7 +45,7 @@ abstract class Animator<T> {
             return;
         }
 
-        float dt = (float)(now - startTime);
+        float dt = now - startTime;
         stepAnimation(dt/DURATION);
 
         if (Build.VERSION.SDK_INT >= 16) {
